refactor: entity 패키지로 이동
