+++ src/main/java/com/aicodinator/backend/domain/community/domain/entity/Post.java
@@ -1,6 +1,6 @@
 package com.aicodinator.backend.domain.community.domain.entity;
 
-import com.aicodinator.backend.domain.region.domain.Region;
+import com.aicodinator.backend.domain.region.domain.entity.Region;
 import com.aicodinator.backend.domain.user.domain.entity.User;
 import com.aicodinator.backend.global.entity.BaseEntity;
 import jakarta.persistence.*;
